+++ CHANGELOG.md
@@ -7,6 +7,7 @@ most recent version is listed first.
 - Enforce `naz` message protocol in code: https://github.com/komuw/naz/pull/179
 - make make `naz.Client.logger` to be a python `logging.Handler`: https://github.com/komuw/naz/pull/181
 - carry out fuzzy testing: https://github.com/komuw/naz/pull/178
+- re-run benchmarks: https://github.com/komuw/naz/pull/182
 
 
 ## **version:** v0.7.3